# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SecureAI - Biometric Security Demo API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uploads:
    get:
      operationId: listUploads
      tags: [uploads]
      summary: List the current user's uploaded files (metadata only, no content)
      responses:
        "200":
          description: Upload metadata list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UploadMeta"

    post:
      operationId: createUpload
      tags: [uploads]
      summary: Encrypt and store an uploaded text, image, or video file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadInput"
      responses:
        "201":
          description: File encrypted and stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadMeta"
        "400":
          description: Bad request (unsupported type or too large)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /uploads/{id}:
    get:
      operationId: getUpload
      tags: [uploads]
      summary: Decrypt and return a file's content (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
      responses:
        "200":
          description: Decrypted file content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadContent"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Upload not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteUpload
      tags: [uploads]
      summary: Delete an uploaded file (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 2147483647
      responses:
        "204":
          description: Upload deleted
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Upload not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UploadMeta:
      type: object
      required: [id, userId, fileName, mimeType, fileType, sizeBytes, createdAt, contentSource, trainingEligible]
      properties:
        id:
          type: number
        userId:
          type: number
        fileName:
          type: string
        mimeType:
          type: string
        fileType:
          type: string
          enum: [image, video, text, audio]
        sizeBytes:
          type: number
        createdAt:
          type: string
        contentSource:
          description: >-
            Declared origin of the file's content, per Team 2's Data Source
            Acceptability Matrix. "unspecified" is the fail-closed default for
            an upload that never declared one.
          type: string
          enum: [own_work, third_party_individual, published_work, social_media, incidental_third_party_ip, unspecified]
        trainingEligible:
          description: >-
            Whether the matrix admits this file into a training corpus, given
            its source and file type together. Returned so the consequence of
            a provenance declaration is visible to the uploader rather than
            only enforced server-side.
          type: boolean
        trainingExclusionReason:
          description: Present only when trainingEligible is false; the matrix's own reasoning.
          type: string
    UploadContent:
      type: object
      required: [id, userId, fileName, mimeType, fileType, sizeBytes, createdAt, dataBase64]
      properties:
        id:
          type: number
        userId:
          type: number
        fileName:
          type: string
        mimeType:
          type: string
        fileType:
          type: string
          enum: [image, video, text, audio]
        sizeBytes:
          type: number
        createdAt:
          type: string
        dataBase64:
          description: Decrypted file content, base64-encoded
          type: string
    UploadInput:
      type: object
      required: [fileName, mimeType, dataBase64]
      properties:
        fileName:
          type: string
          minLength: 1
          maxLength: 255
        mimeType:
          type: string
          minLength: 1
        dataBase64:
          description: Raw file bytes, base64-encoded
          type: string
          minLength: 1
        contentSource:
          description: >-
            Declared origin of the content, per Team 2's Data Source
            Acceptability Matrix. Optional: omitting it stores the upload as
            "unspecified", which keeps the file fully usable by its owner but
            excludes it from every training corpus until a source is declared.
          type: string
          enum: [own_work, third_party_individual, published_work, social_media, incidental_third_party_ip, unspecified]